refactor(SrpInputGrid): extract seed phrase change logic into a separate utility function
This commit refactors the seed phrase handling logic in the SrpInputGrid component by extracting it into a new utility function, `applySeedPhraseChangeAtIndex`, located in `srpInputGridLogic.ts`. This change improves code readability and maintainability by separating concerns, allowing the main component to focus on rendering while delegating the logic to the utility function. The new function handles the seed phrase updates, including validation and state management, previously embedded within the component.
